required plugin gets enabled on action, but not persisted over restart #59
move platform tests to the rest of the tests (as agreed in order to simplify the structure)
