--- src/test/java/fi/aalto/cs/intellij/actions/InstallPluginsNotificationActionFTest.java
@@ -1,17 +1,0 @@
-package fi.aalto.cs.intellij.actions;
-
-import fi.aalto.cs.intellij.PluginsTestHelper;
-import org.junit.Test;
-
-
-public class InstallPluginsNotificationActionFTest extends PluginsTestHelper {
-
-  @Test
-  public void testProposeRestart() {
-    assertThrows(
-        RuntimeException.class,
-        "Restart IntelliJ IDEA to apply changes in plugins?",
-        InstallPluginsNotificationAction::proposeRestart
-    );
-  }
-}
