+++ src/test/java/fi/aalto/cs/intellij/actions/InstallPluginsNotificationActionTest.java
@@ -39,4 +39,13 @@ public class InstallPluginsNotificationActionTest extends PluginsTestHelper {
     assertEquals("Restart proposing functionality should have been called once.", 1,
         numberOfCallsOfRestartProposal.get());
   }
+
+  @Test
+  public void testProposeRestart() {
+    assertThrows(
+        RuntimeException.class,
+        "Restart IntelliJ IDEA to apply changes in plugins?",
+        InstallPluginsNotificationAction::proposeRestart
+    );
+  }
 }
